Move network tools to xelnaga.
