+++ tests/SDK.Tests/Aiursoft.SDK.Tests.csproj
@@ -5,12 +5,6 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.0" />
-		<PackageReference Include="MSTest.TestAdapter" Version="2.1.2" />
-		<PackageReference Include="MSTest.TestFramework" Version="2.1.2" />
-	</ItemGroup>
-
-	<ItemGroup>
 	  <ProjectReference Include="..\..\src\Infrastructure\Developer.SDK\Aiursoft.Developer.SDK.csproj" />
 	  <ProjectReference Include="..\..\src\SDK\SDK\Aiursoft.SDK.csproj" />
 	</ItemGroup>
