+++ tests/functionalTests/Aiursoft.Archon.Tests/BasicTest.cs
@@ -5,6 +5,7 @@ using Aiursoft.Archon.SDK.Services.ToArchonServer;
 using Aiursoft.Handler.Exceptions;
 using Aiursoft.Handler.Models;
 using Aiursoft.SDK.Tests;
+using Aiursoft.XelNaga.Tools;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
