+++ tests/functionalTests/Aiursoft.Observer.Tests/BasicTest.cs
@@ -5,6 +5,7 @@ using Aiursoft.Observer.SDK;
 using Aiursoft.Observer.SDK.Services.ToObserverServer;
 using Aiursoft.SDK;
 using Aiursoft.SDK.Tests;
+using Aiursoft.XelNaga.Tools;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
